Obtain OS info and IP address when guest is not available
Some cloud images have qemu guest agent built in, some do not. So far
I happened to work with images that do, which allowed me to rely on
`domain.guestInfo()`.

The unfortunate byproduct of blind calling `domain.guestInfo()` is
warning polluting STDERR when guest is not available.

With that patch:
- use domain XML to see if guest is available and connected. Use
  `domain.guestInfo()` only when it is.
- otherwise, use domain XML to obtain os_id and os_version_id
- also use domain XML to obtain interface mac addresses, as well as DHCP
  leases. We can join both on mac address and obtain machine IP address
  this way.